feat: add connect circles component (wip)
