+++ circles-app/src/lib/pages/SafeList.svelte
@@ -6,9 +6,10 @@
   import { wallet } from '$lib/stores/wallet';
   import ConnectCircles from '$lib/components/ConnectCircles.svelte';
   import Avatar from '$lib/components/avatar/Avatar.svelte';
+  import { fetchGroupsByOwner } from '$lib/utils/groups';
 
   let safes: string[] = [];
-  let groupCounts: Record<string, number> = {};
+  let groupsByAddress: Record<string, string[]>;
 
   const getSafesByOwnerApiEndpoint = (checksumOwnerAddress: string): string =>
     `https://safe-transaction-gnosis-chain.safe.global/api/v1/owners/${checksumOwnerAddress}/safes/`;
@@ -23,40 +24,6 @@
     const safesByOwner = await safesByOwnerResult.json();
 
     return safesByOwner.safes ?? [];
-  }
-
-  async function fetchGroupsByOwner(ownerAddress: string): Promise<any> {
-    const payload = {
-      jsonrpc: '2.0',
-      id: 1,
-      method: 'circles_query',
-      params: [
-        {
-          Namespace: 'CrcV2',
-          Table: 'CMGroupCreated',
-          Columns: ['proxy'],
-          Filter: [
-            {
-              Type: 'FilterPredicate',
-              FilterType: 'Equals',
-              Column: 'owner',
-              Value: ownerAddress.toLowerCase(),
-            },
-          ],
-          Order: [],
-          Limit: 10,
-        },
-      ],
-    };
-
-    const response = await fetch('https://rpc.circlesubi.network/', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(payload),
-    });
-
-    const data = await response.json();
-    return data.result.rows;
   }
 
   onMount(loadSafesAndGroups);
@@ -75,7 +42,8 @@
 
     const groupFetchPromises = safes.map(async (safe) => {
       const groups = await fetchGroupsByOwner(safe);
-      groupCounts = { ...groupCounts, [safe]: groups.length };
+      console.log(groups);
+      groupsByAddress = { ...groupsByAddress, [safe]: groups.flat() };
     });
 
     await Promise.all(groupFetchPromises);
@@ -84,6 +52,11 @@
   //
   // Connects the wallet and initializes the Circles SDK.
   //
+
+  // bottomInfo={shortenAddress(item.toLowerCase()) +
+  //       (groupsByAddress[item].length > 0
+  //         ? ' - owner ' + groupsByAddress[item].length + ' groups'
+  //         : '')}
 </script>
 
 {#each safes ?? [] as item (item)}
@@ -92,12 +65,9 @@
       address={item.toLowerCase()}
       clickable={false}
       view="horizontal"
-      bottomInfo={shortenAddress(item.toLowerCase()) +
-        (groupCounts[item] > 0
-          ? ' - owner ' + groupCounts[item] + ' groups'
-          : '')}
+      
     />
-      </ConnectCircles>
+  </ConnectCircles>
 {/each}
 {#if (safes ?? []).length === 0}
   <div class="text-center">
